% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotateDmrs}
\alias{annotateDmrs}
\title{Annotate DMRs with genome annotations}
\usage{
annotateDmrs(
  gr,
  annots_beds = NULL,
  annots_names = NULL,
  bsseq_paths = NULL,
  hg38_features = c("hg38_genes_promoters", "hg38_genes_intergenic", "hg38_cpg_islands",
    "hg38_basicgenes")
)
}
\arguments{
\item{gr}{character or GRanges object. The DMRs to annotate. If a character, assumed to be a path to a .rda file with a GRanges object named 'filtered_dmrs'.}

\item{annots_beds}{character vector of file paths to custom annotation BED files.}

\item{annots_names}{character vector of names for custom annotations. If NULL, the names will be the base names of the BED files.}

\item{bsseq_paths}{character vector of file paths to BSseq objects.}

\item{hg38_features}{character vector of genome annotation features to include. Default: c("hg38_genes_promoters", "hg38_genes_intergenic", "hg38_cpg_islands", "hg38_basicgenes").}
}
\value{
GRanges object with annotations.
}
\description{
Annotate DMRs with genome annotations, including custom annotations, and methylation values from user-provided BSseq objects.
}
