% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bed2bsseq}
\alias{bed2bsseq}
\title{bed2bsseq}
\usage{
bed2bsseq(file, samplename = basename(file))
}
\arguments{
\item{file}{A simplfied bed file with methylation: chr\tstart\tend\tmethylation\coverage}

\item{samplename}{basename(file)}
}
\value{
a bsseq object
}
\description{
bed2bsseq
}
\examples{
bed2bsseq(file=system.file("extdata", "test.bed.gz", package = "dhsrmethtools"),samplename="test")
}
