% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{callDmrs}
\alias{callDmrs}
\title{callDmrs}
\usage{
callDmrs(
  sample_dir,
  normal_dir,
  sample_name = NULL,
  normal_name = NULL,
  cpg_smoothing = TRUE,
  ncores = 1,
  dmr_delta = 0.35,
  dmr_p_threshold = 0.05,
  min.cpg.coverage = 2,
  min.cpgs = 5,
  min.length = 100,
  output = NULL
)
}
\arguments{
\item{sample_dir}{Directory containing the sample BSseq HDF5Array.}

\item{normal_dir}{Directory containing the normal BSseq HDF5Array.}

\item{sample_name}{Name of the sample column in the sample BSseq object. If NULL, it will be inferred assuming a .combined pattern.}

\item{normal_name}{Name of the normal column in the normal BSseq object. If NULL, it will be inferred assuming a .combined pattern.}

\item{cpg_smoothing}{Logical, whether to perform smoothing. Default is TRUE.}

\item{ncores}{Integer, number of cores to use. Default is 1.}

\item{dmr_delta}{Numeric, delta value for callDMR function. Default is 0.35.}

\item{dmr_p_threshold}{Numeric, p-value threshold for callDMR function. Default is 0.05.}

\item{min.cpg.coverage}{Integer, minimum coverage for a CpG site to be considered. Default is 2.}

\item{min.cpgs}{Integer, minimum number of CpG sites in a region to be considered. Default is 5.}

\item{min.length}{Integer, minimum length of a DMR region to be considered. Default is 100.}

\item{output}{Character, path to save the output. If NULL, the function returns the DMR object.
The file extension determines the output format:
- "rda": Saves the DMRs as an R data file (.rda).
- "bed": Saves the DMRs as a BED file (.bed) with no header, including columns "chr", "start", "end", "diff.Methy".
- "tsv": Saves the DMRs as a tab-separated values file (.tsv) with a header.
If the file extension is not one of these, the function stops and returns an error.}
}
\value{
A message indicating the success or failure of the operation, or the DMR object if output is NULL.
}
\description{
Perform differential methylation analysis on a BSseq object for a sample compared to a normal and save or return the DMRs.
}
\examples{
callDmrs(
  sample_dir = "path/to/sample_dir",
  normal_dir = "path/to/normal_dir",
  output = "path/to/output/DMRs.sample_name.vs.normal_name.rda"
)
}
