% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{callHapDmrs}
\alias{callHapDmrs}
\title{callHapDmrs}
\usage{
callHapDmrs(
  bs,
  group1 = sampleNames(bs)[grep("\\\\.hap1$", sampleNames(bs))],
  group2 = sampleNames(bs)[grep("\\\\.hap2$", sampleNames(bs))],
  cpg_smoothing = TRUE,
  ncores = 1,
  dmr_delta = 0.35,
  dmr_p_threshold = 0.05,
  min.cpg.coverage = 2,
  min.cpgs = 5,
  min.length = 100,
  output = NULL
)
}
\arguments{
\item{bs}{BSseq object or path to a BSseq HDF5Array directory.
If a path is provided, the BSseq object will be loaded from the specified directory.}

\item{group1}{Character vector, sample names for group 1. Default is samples with suffix ".hap1".}

\item{group2}{Character vector, sample names for group 2. Default is samples with suffix ".hap2".}

\item{cpg_smoothing}{Logical, whether to perform smoothing. Default is TRUE.}

\item{ncores}{Integer, number of cores to use. Default is 1.}

\item{dmr_delta}{Numeric, delta value for callDMR function. Default is 0.35.}

\item{dmr_p_threshold}{Numeric, p-value threshold for callDMR function. Default is 0.05.}

\item{min.cpg.coverage}{Integer, minimum coverage for a CpG site to be considered. Default is 2.}

\item{min.cpgs}{Integer, minimum number of CpG sites in a region to be considered. Default is 5.}

\item{min.length}{Integer, minimum length of a DMR region to be considered. Default is 100.}

\item{output}{Character, path to save the output. If NULL, the function returns the DMR object.
The file extension determines the output format:
- "rda": Saves the DMRs as an R data file (.rda).
- "bed": Saves the DMRs as a BED file (.bed) with no header, including columns "chr", "start", "end", "diff.Methy".
- "tsv": Saves the DMRs as a tab-separated values file (.tsv) with a header.
If the file extension is not one of these, the function stops and returns an error.}
}
\value{
A message indicating the success or failure of the operation, or the DMR object if output is NULL.
}
\description{
Perform differential methylation analysis on a BSseq object for two haplotypes in the same sample and save or return the DMRs.
}
\examples{
callHapDmrs(
  bs = "path/to/bsseq_object",
  output = "path/to/output/DMRs.sample_name.hap.rda"
)
}
