% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterDmrs}
\alias{filterDmrs}
\title{filterDmrs
Filter Differentially Methylated Regions}
\usage{
filterDmrs(
  gr,
  bs,
  normal.bs = NULL,
  min.diff = 0.4,
  min.cpgs = 5,
  min.length = 100,
  excludeRegions,
  filter.sex.chromosomes = TRUE
)
}
\arguments{
\item{gr}{A GRanges object containing the DMRs to be filtered.}

\item{bs}{A BSseq object containing methylation data for the sample.}

\item{normal.bs}{A BSseq object containing methylation data for the normal sample (optional).}

\item{min.diff}{Numeric value specifying the minimum methylation difference to consider (default: 0.4).}

\item{min.cpgs}{Integer specifying the minimum number of CpGs required in a region (default: 5).}

\item{min.length}{Integer specifying the minimum length of a region in base pairs (default: 100).}

\item{excludeRegions}{A GRanges object specifying regions to exclude from the analysis.}

\item{filter.sex.chromosomes}{Logical indicating whether to filter out sex chromosomes (default: TRUE).}
}
\value{
A filtered GRanges object containing the DMRs that meet the specified criteria.
}
\description{
This function filters differentially methylated regions (DMRs) based on various criteria,
comparing methylation levels between a sample and optionally a normal sample.
}
\details{
The function performs the following main steps:
1. Checks input validity and prepares data.
2. Optionally removes sex chromosomes.
3. Excludes specified regions from the analysis.
4. Filters regions based on various criteria (number of CpGs, length, coverage, methylation difference).
5. Classifies regions as hypomethylated or hypermethylated compared to the normal sample.
}
\examples{
# Basic usage:
filtered_dmrs <- filterDmrs(gr, bs, normal.bs, excludeRegions = excluded_regions)

# With custom parameters:
filtered_dmrs <- filterDmrs(gr, bs, normal.bs, min.diff = 0.3, min.cpgs = 10,
                            excludeRegions = excluded_regions,
                            filter.sex.chromosomes = FALSE)

}
