% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hapBSseq}
\alias{hapBSseq}
\title{hapBSseq}
\usage{
hapBSseq(
  hap1_file,
  hap2_file,
  combined_file,
  sample_name = NULL,
  platform,
  results_path = NULL,
  replace = FALSE
)
}
\arguments{
\item{hap1_file}{Path to the haplotype 1 bedmethyl file.}

\item{hap2_file}{Path to the haplotype 2 bedmethyl file.}

\item{combined_file}{Path to the combined bedmethyl file.}

\item{sample_name}{Sample name. Default is derived from the combined file basename.}

\item{platform}{Platform type ('ont' or 'hifi').}

\item{results_path}{Base path to save the BSseq results. If NULL, the function returns the BSseq object.}

\item{replace}{Logical, whether to replace the existing directory. Default is FALSE.}
}
\value{
A message indicating the success or failure of the operation.
}
\description{
Create a BSseq object from three input files (hap1, hap2, combined) based on the platform type.
}
\examples{
hapBSseq(
  hap1_file = "path/to/hap1_file",
  hap2_file = "path/to/hap2_file",
  combined_file = "path/to/combined_file",
  sample_name = "sample1",
  platform = "ont",
  results_path = "path/to/results",
  replace = FALSE
)
}
