% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inferGRanges}
\alias{inferGRanges}
\title{inferGRanges
Infer GRanges Object from Various Inputs}
\usage{
inferGRanges(input, chr_col = NULL, start_col = NULL, end_col = NULL)
}
\arguments{
\item{input}{Either a file path (.bed, .txt, .tsv, .Rda) or a data frame.}

\item{chr_col}{Character string specifying the column name for chromosome information in the data frame (optional).}

\item{start_col}{Character string specifying the column name for start position information in the data frame (optional).}

\item{end_col}{Character string specifying the column name for end position information in the data frame (optional).}
}
\value{
A GRanges object representing the genomic regions.
}
\description{
This function infers a GRanges object from a file or data frame, providing flexibility
in how genomic regions are specified.
}
\details{
The function can handle various input types:
1. Data frame: Directly converts to GRanges, using specified or default column names.
2. BED file: Reads and converts to GRanges, assuming standard BED format.
3. Text or TSV file: Reads as a table and converts to GRanges.
4. RDA file: Loads an R object named 'dmrs' and converts to GRanges.

If column names are not specified, the function attempts to use standard names
("chr", "start", "end"). For BED files, it handles additional standard BED columns if present.
}
\examples{
# From a data frame
df <- data.frame(chr = c("chr1", "chr2"), start = c(100, 200), end = c(200, 300))
gr <- inferGRanges(df)

# From a BED file
gr <- inferGRanges("path/to/file.bed")

# From a text file with custom column names
gr <- inferGRanges("path/to/file.txt", chr_col = "chromosome", start_col = "begin", end_col = "stop")

}
