% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{modkit2bsseq}
\alias{modkit2bsseq}
\title{modkit2bsseq}
\usage{
modkit2bsseq(file, samplename = basename(file))
}
\arguments{
\item{file}{An ONT bedmethyl file}

\item{samplename}{basename(file)}
}
\value{
a bsseq object
}
\description{
modkit2bsseq
}
\examples{
modkit2bsseq(file=system.file("extdata", "test.basemods.bedmethyl.gz", package = "dhsrmethtools"),samplename="test")
}
