% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parseBamStats}
\alias{parseBamStats}
\title{Parse bamstats output file}
\usage{
parseBamStats(statsfile)
}
\arguments{
\item{statsfile}{path to bamstats output file}
}
\value{
a list with basic stats and histogram of read lengths
}
\description{
Parse bamstats output file
}
