% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pbcpg2bsseq}
\alias{pbcpg2bsseq}
\title{pbcpg2bsseq}
\usage{
pbcpg2bsseq(file, samplename = basename(file))
}
\arguments{
\item{file}{A pb-CpG tools bedmethyl file}

\item{samplename}{basename(file)}
}
\value{
a bsseq object
}
\description{
pbcpg2bsseq
}
\examples{
pbcpg2bsseq(file=system.file("extdata", "test.pbcpg.bedmethyl.gz", package = "dhsrmethtools"),samplename="test")
}
